add change id handler

+++ src/main/java/dev/dov/image_storage_service/image/ImageServiceImp.java
@@ -1,13 +1,13 @@
 package dev.dov.image_storage_service.image;
 
+import dev.dov.image_storage_service.image.enums.ImagesRequestType;
 import dev.dov.image_storage_service.image.interfaces.ImageService;
 import io.minio.*;
 import io.minio.errors.*;
 import io.minio.http.Method;
-import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
+import io.minio.messages.DeleteObject;
+import io.minio.messages.Item;
 import lombok.SneakyThrows;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
@@ -17,11 +17,15 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
 @Service
 public class ImageServiceImp implements ImageService {
 
-    @Value("${garbage.bucket.name}")
+    @Value("${garbage.bucket-name}")
     private String bucketName;
 
     private final MinioClient minioClient;
@@ -62,23 +66,52 @@ public class ImageServiceImp implements ImageService {
     }
 
     @Override
-    public void addImage(String filename, InputStream inputStream) {
-
-        try(BufferedInputStream bis = new BufferedInputStream(inputStream)) {
+    public void addImage(String filename, InputStream inputStream, String contentType, boolean overwrite) {
+        try (BufferedInputStream bis = new BufferedInputStream(inputStream)) {
 
             ensureBucketExists();
 
+            if (!overwrite && isSuchImageAlreadyExist(filename)) {
+                throw new IllegalStateException("Зображення '" + filename + "' вже існує.");
+            }
+
             minioClient.putObject(
                     PutObjectArgs.builder()
-                    .bucket(bucketName)
+                            .bucket(bucketName)
                             .object(filename)
+                            .contentType(contentType)
                             .stream(bis, bis.available(), 5 * 1024 * 1024)
-                            .build());
+                            .build()
+            );
 
         } catch (Exception e) {
             e.printStackTrace();
         }
+    }
 
+    private boolean isSuchImageAlreadyExist(String filename) throws InsufficientDataException, InternalException, InvalidKeyException, InvalidResponseException, IOException, NoSuchAlgorithmException, ServerException, XmlParserException, ErrorResponseException {
+
+        try {
+            minioClient.statObject(
+                    StatObjectArgs.builder()
+                            .bucket(bucketName)
+                            .object(filename)
+                            .build()
+            );
+
+            return true;
+
+        } catch (io.minio.errors.ErrorResponseException e) {
+
+            if ("NoSuchKey".equals(e.errorResponse().code())) {
+                return false;
+            }
+
+            throw new RuntimeException("Помилка при перевірці існування зображення: " + e.getMessage(), e);
+
+        } catch (Exception e) {
+            throw new RuntimeException("Помилка при перевірці існування зображення: " + e.getMessage(), e);
+        }
     }
 
 
@@ -92,7 +125,6 @@ public class ImageServiceImp implements ImageService {
     public String getPresignedObjectUrl(String filename) {
 
         ensureBucketExists();
-
         return minioClient.getPresignedObjectUrl(
                 GetPresignedObjectUrlArgs.builder()
                         .bucket(bucketName)
@@ -115,4 +147,119 @@ public class ImageServiceImp implements ImageService {
             );
         }
     }
+
+    @Override
+    @SneakyThrows
+    public void renameFilesByPrefix(String oldPrefix, String newPrefix) {
+        ensureBucketExists();
+
+        Iterable<Result<Item>> results = minioClient.listObjects(
+                ListObjectsArgs.builder()
+                        .bucket(bucketName)
+                        .prefix(oldPrefix)
+                        .recursive(true)
+                        .build()
+        );
+
+        for (Result<Item> result : results) {
+            Item item = result.get();
+            String oldName = item.objectName();
+            String newName = oldName.replaceFirst("^" + oldPrefix, newPrefix);
+
+            minioClient.copyObject(
+                    CopyObjectArgs.builder()
+                            .source(CopySource.builder()
+                                    .bucket(bucketName)
+                                    .object(oldName)
+                                    .build())
+                            .bucket(bucketName)
+                            .object(newName)
+                            .build()
+            );
+
+            minioClient.removeObject(
+                    RemoveObjectArgs.builder()
+                            .bucket(bucketName)
+                            .object(oldName)
+                            .build()
+            );
+        }
+    }
+
+    @Override
+    @SneakyThrows(Exception.class)
+    public Map<String, String> getPresignedUrlsByPrefixAndType(String locationId, ImagesRequestType type) {
+
+        ensureBucketExists();
+
+        Map<String, String> urlsMap = new HashMap<>();
+
+        String imageName = "def";
+
+        switch (type) {
+            case LOCATION -> imageName = "loc-".concat(locationId).concat("_");
+            case CHECK -> imageName = "check-".concat(locationId).concat("_");
+        }
+
+        Iterable<Result<Item>> results = minioClient.listObjects(
+                ListObjectsArgs.builder()
+                        .bucket(bucketName)
+                        .prefix(imageName)
+                        .recursive(true)
+                        .build()
+        );
+
+        for (Result<Item> result : results) {
+            Item item = result.get();
+            String objectName = item.objectName();
+            String url = minioClient.getPresignedObjectUrl(
+                    GetPresignedObjectUrlArgs.builder()
+                            .bucket(bucketName)
+                            .object(objectName)
+                            .method(Method.GET)
+                            .build()
+            );
+            String[] imageIdParts = objectName.split("_");
+
+            switch (imageIdParts.length) {
+                case 2 -> urlsMap.put(imageIdParts[imageIdParts.length-1], url);
+                case 3 -> urlsMap.put(imageIdParts[1], url);
+            }
+
+        }
+
+        return urlsMap;
+    }
+
+    @Override
+    @SneakyThrows(Exception.class)
+    public void deleteImageByPrefix(String prefix) {
+        ensureBucketExists();
+
+        Iterable<Result<Item>> results = minioClient.listObjects(
+                ListObjectsArgs.builder()
+                        .bucket(bucketName)
+                        .prefix(prefix)
+                        .recursive(true)
+                        .build()
+        );
+
+        List<DeleteObject> objectsToDelete = new ArrayList<>();
+
+        for (Result<Item> result : results) {
+            String objectName = result.get().objectName();
+            objectsToDelete.add(new DeleteObject(objectName));
+        }
+
+        if (!objectsToDelete.isEmpty()) {
+            minioClient.removeObjects(
+                    RemoveObjectsArgs.builder()
+                            .bucket(bucketName)
+                            .objects(objectsToDelete)
+                            .build()
+            );
+        }
+    }
+
+
 }
